// libs/modules/sps-website-builder/relations/navbar-blocks-to-buttons/frontend/component/variants/sps-lite/select-right/src/lib/interface.ts
import { IRelation } from "@sps/sps-website-builder-relations-navbar-blocks-to-buttons-contracts";
import { IRelation as IRelationExtended } from "@sps/sps-website-builder-relations-navbar-blocks-to-buttons-contracts-extended";

export const variant = "select-right" as const;

export interface IComponentBase {
  showSkeletons?: boolean;
  isServer: boolean;
}

export interface IComponentProps extends IComponentBase {
  variant: typeof variant;
  navbarBlockId?: string;
  data?: IRelation;
}

export interface IComponentPropsExtended extends IComponentProps {
  data?: IRelationExtended;
}